# Defer nvm plugin from loading until it is needed